requirements.txt
opencv-python==4.4.0
numpy==1.19.2
Pillow==8.0.0
pywin32==228
pytesseract==0.3.6